pyproject.toml
[project]
name = "cto-tool"
version = "1.0.0"
description = ""
authors = []
requires-python = "~=3.12.3"
readme = "README.md"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.14",
    "aiosignal==1.4.0",
    "annotated-types==0.7.0",
    "anthropic==0.59.0",
    "anyio==4.9.0",
    "attrs==24.3.0",
    "azure-devops>=7.1.0b4,<8",
    "beautifulsoup4>=4.9.1,<5",
    "boto3==1.36.21",
    "botocore==1.36.21",
    "certifi==2024.12.14",
    "chardet>=5.2.0,<6",
    "charset-normalizer==3.4.1",
    "contourpy==1.3.1",
    "cycler==0.12.1",
    "dataclasses-json==0.6.7",
    "defusedxml==0.7.1",
    "distro==1.9.0",
    "django>=4.2.11,<5",
    "django-allauth[mfa]>=65.1.0,<66",
    "django-auditlog>=3.0.0,<4",
    "django-admin-env-notice>=1.0.1",
    "django-environ>=0.11.2,<0.12",
    "django-extensions>=3.2.3,<4",
    "django-filter~=24.2",
    "django-hijack>=3.4.1,<4",
    "django-imagekit>=5.0.0,<6",
    "django-import-export>=3.3.5,<4",
    "django-inline-static>=0.1.0,<0.2",
    "django-multiselectfield>=0.1.13,<0.2",
    "django-oauth-toolkit>=2.3.0,<3",
    "django-soft-delete>=0.9.21,<0.10",
    "django-storages>=1.14.4,<2",
    "django-timescaledb>=0.2.13,<0.3",
    "django-treebeard>=4.7.1,<5",
    "django-widget-tweaks>=1.4.12,<2",
    "djangorestframework>=3.15.1,<4",
    "fonttools==4.59.0",
    "frozenlist==1.7.0",
    "gitpython>=3.1.44,<4",
    "gorilla>=0.4.0,<0.5",
    "greenlet==3.2.3",
    "h11==0.16.0",
    "hashids>=1.3.1,<2",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "httpx-sse==0.4.0",
    "idna==3.10",
    "ipdb>=0.13.9,<0.14",
    "ipython~=7.28",
    "jiter==0.10.0",
    "jmespath==1.0.1",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jwt>=1.3.1,<2",
    "kiwisolver==1.4.8",
    "langchain==0.3.14",
    "langchain-anthropic==0.3.1",
    "langchain-aws==0.2.12",
    "langchain-community==0.3.14",
    "langchain-core==0.3.29",
    "langchain-openai==0.2.14",
    "langchain-text-splitters==0.3.5",
    "langsmith==0.2.10",
    "mailchimp-marketing>=3.0.80,<4",
    "markdown>=3.5.2,<4",
    "marshmallow==3.26.1",
    "matplotlib==3.10.0",
    "msrest>=0.7.1,<0.8",
    "multidict==6.6.3",
    "mypy-extensions==1.1.0",
    "numpy==1.26.4",
    "oauthlib>=3.2.2,<4",
    "oauth2-provider>=0.0,<0.1",
    "openai==1.59.4",
    "opentelemetry-distro>=0.53b1",
    "opentelemetry-exporter-otlp>=1.32.1",
    "opentelemetry-instrumentation>=0.53b1",
    "opentelemetry-instrumentation-aiohttp-client>=0.53b1",
    "opentelemetry-instrumentation-aiohttp-server>=0.53b1",
    "opentelemetry-instrumentation-asgi>=0.53b1",
    "opentelemetry-instrumentation-asyncio>=0.53b1",
    "opentelemetry-instrumentation-boto3sqs>=0.53b1",
    "opentelemetry-instrumentation-botocore>=0.53b1",
    "opentelemetry-instrumentation-click>=0.53b1",
    "opentelemetry-instrumentation-dbapi>=0.53b1",
    "opentelemetry-instrumentation-django>=0.53b1",
    "opentelemetry-instrumentation-grpc>=0.53b1",
    "opentelemetry-instrumentation-httpx>=0.53b1",
    "opentelemetry-instrumentation-jinja2>=0.53b1",
    "opentelemetry-instrumentation-langchain>=0.40.9,<0.41",
    "opentelemetry-instrumentation-logging>=0.53b1",
    "opentelemetry-instrumentation-openai-v2~=2.1b0",
    "opentelemetry-instrumentation-psycopg>=0.53b1",
    "opentelemetry-instrumentation-requests>=0.53b1",
    "opentelemetry-instrumentation-sqlalchemy>=0.53b1",
    "opentelemetry-instrumentation-sqlite3>=0.53b1",
    "opentelemetry-instrumentation-threading>=0.53b1",
    "opentelemetry-instrumentation-tortoiseorm>=0.53b1",
    "opentelemetry-instrumentation-urllib>=0.53b1",
    "opentelemetry-instrumentation-urllib3>=0.53b1",
    "opentelemetry-instrumentation-wsgi>=0.53b1",
    "openpyxl>=3.1.2,<4",
    "orjson==3.11.0",
    "otel-extensions>=1.1.0,<2",
    "packaging==24.2",
    "pandas==2.3.0",
    "parameterized>=0.9.0,<0.10",
    "pillow==11.3.0",
    "posthog>=3.0.2,<4",
    "propcache==0.3.2",
    "psycopg>=3.1.17,<4",
    "pydantic==2.10.4",
    "pydantic-core==2.27.2",
    "pydantic-settings==2.10.1",
    "pyparsing==3.2.1",
    "pypdf==5.9.0",
    "pysh>=3.3.1,<4",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "pytz==2024.2",
    "pyyaml==6.0.2",
    "regex==2024.11.6",
    "requests==2.32.4",
    "requests-toolbelt==1.0.0",
    "s3transfer==0.11.2",
    "seaborn==0.13.2",
    "sentry-sdk[django]>=2.27.0,<3",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.36",
    "statsmodels>=0.14.4,<0.15",
    "tblib>=3.0.0,<4",
    "tenacity==9.1.2",
    "textdistance>=4.6.3,<5",
    "thefuzz>=0.22.1,<0.23",
    "tiktoken==0.9.0",
    "timeout-decorator==0.5.0",
    "tqdm==4.67.1",
    "typing==3.7.4.3",
    "typing-extensions==4.14.0",
    "typing-inspect==0.9.0",
    "tzdata==2024.2",
    "urllib3==1.26.20",
    "yarl==1.20.1",
    "django-cte>=2.0.0",
    "langchain-google-genai>=2.0.9",
    "jira2markdown>=0.5",
]

[dependency-groups]
dev = [
    "black>=24.4.2,<25",
    "django-debug-toolbar>=4.4.6,<5",
    "django-silk>=5.1.0,<6",
    "isort>=5.13.2,<6",
    "pre-commit>=4.2.0,<5",
    "psycopg2-binary>=2.9.10",
    "ruff>=0.11.13,<0.12",
    "setuptools>=79.0.0,<80",
    "vcrpy>=7.0.0,<8",
    "ydata-profiling>=4.16.1,<5",
]

test = [
    "pytest>=8.3.5,<9",
    "faker>=26.0.0,<27",
]

[tool.uv]
package = false
default-groups = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
